requirements.txt
flask
flask-cors
numpy
opencv-python
mxnet
gunicorn
pillow